requirements.txt
requests==2.31.0
dnspython==2.4.2
python-whois==0.8.0
colorama==0.4.6
pyfiglet==0.8.post1
matplotlib==3.7.1
networkx==3.1